backend/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
scikit-learn==0.24.1
scikit-image==0.18.1
pytest==7.1.0
numpy==1.22.3
pandas==1.4.1
pytest-django==4.5.2